+++ gql_exec/lib/gql_exec.dart
@@ -1,0 +1,8 @@
+/// Basis for GraphQL execution layer to support Link and Client.
+library execution;
+
+export "./exec/context.dart";
+export "./exec/error.dart";
+export "./exec/operation.dart";
+export "./exec/request.dart";
+export "./exec/response.dart";