+++ gql_transform_link/test/gql_transform_link_test.dart
@@ -1,6 +1,6 @@
 import "dart:async";
 
-import "package:gql/execution.dart";
+import "package:gql_exec/gql_exec.dart";
 import "package:gql/language.dart";
 import "package:gql_link/gql_link.dart";
 import "package:gql_transform_link/gql_transform_link.dart";