+++ gql_http_link/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.4
+
+- use `package:gql_exec`
+
 ## 0.2.3
 
 - upgrade `package:gql_link` to v0.2.0